+++ .dangerfile/changelog_tests.rb
@@ -41,7 +41,7 @@ def changelog_bad_formatting?(file)
     end
   end
 
-  fail_message = "**#{file}**\nCHANGELOG.md has formatting problems. Please correct the below:\n\n" + fail_message if !fail_message.empty?
+  fail_message = "CHANGELOG.md has formatting problems. Please correct the below:\n\n" + fail_message if !fail_message.empty?
 
   return fail_message.strip if !fail_message.empty?
   return false