+++ security/aws/rds_publicly_accessible/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v4.2.1
+
+- Added default value for parameters that do not require user input
+
 ## v4.2
 
 - Changed internal name of escalation code to ensure "Meta Policy" works as expected